+++ lapce-data/src/keypress/loader.rs
@@ -1,0 +1,175 @@
+use anyhow::{anyhow, Result};
+use indexmap::IndexMap;
+use toml;
+
+use crate::{
+    keypress::{get_modes, KeyMap, KeyPress},
+    state::Mode,
+};
+
+pub struct KeyMapLoader {
+    keymaps: IndexMap<Vec<KeyPress>, Vec<KeyMap>>,
+    command_keymaps: IndexMap<String, Vec<KeyMap>>,
+}
+
+impl KeyMapLoader {
+    pub fn new() -> Self {
+        Self {
+            keymaps: Default::default(),
+            command_keymaps: Default::default(),
+        }
+    }
+
+    pub fn load_from_str<'a>(
+        &'a mut self,
+        s: &str,
+        modal: bool,
+    ) -> Result<&'a mut Self> {
+        let toml_keymaps: toml::Value = toml::from_str(s)?;
+        let toml_keymaps = toml_keymaps
+            .get("keymaps")
+            .and_then(|v| v.as_array())
+            .ok_or_else(|| anyhow!("no keymaps"))?;
+
+        for toml_keymap in toml_keymaps {
+            let keymap = match Self::get_keymap(toml_keymap, modal) {
+                Ok(keymap) => keymap,
+                Err(err) => {
+                    log::error!("Could not parse keymap: {err}");
+                    continue;
+                }
+            };
+
+            let (command, bind) = match keymap.command.strip_prefix('-') {
+                Some(cmd) => (cmd.to_string(), false),
+                None => (keymap.command.clone(), true),
+            };
+
+            let current_keymaps = self.command_keymaps.entry(command).or_default();
+            if bind {
+                current_keymaps.push(keymap.clone());
+                for i in 1..keymap.key.len() + 1 {
+                    let key = keymap.key[..i].to_vec();
+                    self.keymaps.entry(key).or_default().push(keymap.clone());
+                }
+            } else {
+                let is_keymap = |k: &KeyMap| -> bool {
+                    k.when == keymap.when
+                        && k.modes == keymap.modes
+                        && k.key == keymap.key
+                };
+                if let Some(index) = current_keymaps.iter().position(is_keymap) {
+                    current_keymaps.remove(index);
+                }
+                for i in 1..keymap.key.len() + 1 {
+                    if let Some(keymaps) = self.keymaps.get_mut(&keymap.key[..i]) {
+                        if let Some(index) = keymaps.iter().position(is_keymap) {
+                            keymaps.remove(index);
+                        }
+                    }
+                }
+            }
+        }
+
+        Ok(self)
+    }
+
+    pub fn finalize(
+        self,
+    ) -> (
+        IndexMap<Vec<KeyPress>, Vec<KeyMap>>,
+        IndexMap<String, Vec<KeyMap>>,
+    ) {
+        let Self {
+            keymaps: map,
+            command_keymaps: command_map,
+        } = self;
+
+        (map, command_map)
+    }
+
+    fn get_keymap(toml_keymap: &toml::Value, modal: bool) -> Result<KeyMap> {
+        let key = toml_keymap
+            .get("key")
+            .and_then(|v| v.as_str())
+            .ok_or_else(|| anyhow!("no key in keymap"))?;
+
+        let modes = get_modes(toml_keymap);
+        if !modal && !modes.is_empty() && !modes.contains(&Mode::Insert) {
+            return Err(anyhow!(""));
+        }
+
+        Ok(KeyMap {
+            key: KeyPress::parse(key),
+            modes: get_modes(toml_keymap),
+            when: toml_keymap
+                .get("when")
+                .and_then(|w| w.as_str())
+                .map(|w| w.to_string()),
+            command: toml_keymap
+                .get("command")
+                .and_then(|c| c.as_str())
+                .map(|w| w.trim().to_string())
+                .unwrap_or_else(|| "".to_string()),
+        })
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_keymap() {
+        let keymaps = r###"
+keymaps = [
+    { key = "ctrl+w l l", command = "right", when = "n" },
+    { key = "ctrl+w l", command = "right", when = "n" },
+    { key = "ctrl+w h", command = "left", when = "n" },
+    { key = "ctrl+w",   command = "left", when = "n" },
+    { key = "End", command = "line_end", when = "n" },
+    { key = "I", command = "insert_first_non_blank", when = "n" },
+]
+        "###;
+        let mut loader = KeyMapLoader::new();
+        loader.load_from_str(keymaps, true).unwrap();
+
+        let (keymaps, _) = loader.finalize();
+
+        // Lower case modifiers
+        let keypress = KeyPress::parse("ctrl+w");
+        assert_eq!(keymaps.get(&keypress).unwrap().len(), 4);
+
+        let keypress = KeyPress::parse("ctrl+w l");
+        assert_eq!(keymaps.get(&keypress).unwrap().len(), 2);
+
+        let keypress = KeyPress::parse("ctrl+w h");
+        assert_eq!(keymaps.get(&keypress).unwrap().len(), 1);
+
+        let keypress = KeyPress::parse("ctrl+w l l");
+        assert_eq!(keymaps.get(&keypress).unwrap().len(), 1);
+
+        let keypress = KeyPress::parse("end");
+        assert_eq!(keymaps.get(&keypress).unwrap().len(), 1);
+
+        // Upper case modifiers
+        let keypress = KeyPress::parse("Ctrl+w");
+        assert_eq!(keymaps.get(&keypress).unwrap().len(), 4);
+
+        let keypress = KeyPress::parse("Ctrl+w l");
+        assert_eq!(keymaps.get(&keypress).unwrap().len(), 2);
+
+        let keypress = KeyPress::parse("Ctrl+w h");
+        assert_eq!(keymaps.get(&keypress).unwrap().len(), 1);
+
+        let keypress = KeyPress::parse("Ctrl+w l l");
+        assert_eq!(keymaps.get(&keypress).unwrap().len(), 1);
+
+        let keypress = KeyPress::parse("End");
+        assert_eq!(keymaps.get(&keypress).unwrap().len(), 1);
+
+        // No modifier
+        let keypress = KeyPress::parse("I");
+        assert_eq!(keymaps.get(&keypress).unwrap().len(), 1);
+    }
+}
