+++ lapce-data/src/keypress/mod.rs
@@ -1,9 +1,8 @@
-use std::fmt::Display;
 use std::path::PathBuf;
 use std::str::FromStr;
 use std::sync::Arc;
 
-use anyhow::{anyhow, Result};
+use anyhow::Result;
 use druid::piet::{PietTextLayout, Text, TextLayout, TextLayoutBuilder};
 use druid::{Command, KbKey};
 use druid::{
@@ -16,12 +15,18 @@ use indexmap::IndexMap;
 use itertools::Itertools;
 use toml;
 
+mod keypress;
+mod loader;
+
 use crate::command::{
     lapce_internal_commands, CommandExecuted, CommandTarget, LapceCommandNew,
     LapceUICommand, LAPCE_NEW_COMMAND, LAPCE_UI_COMMAND,
 };
 use crate::config::{Config, LapceTheme};
+use crate::keypress::loader::KeyMapLoader;
 use crate::{command::LapceCommand, state::Mode};
+
+pub use keypress::KeyPress;
 
 const DEFAULT_KEYMAPS_COMMON: &str =
     include_str!("../../../defaults/keymaps-common.toml");
@@ -38,114 +43,6 @@ enum KeymapMatch {
     Multiple(Vec<String>),
     Prefix,
     None,
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Hash)]
-pub struct KeyPress {
-    pub key: druid::KbKey,
-    pub mods: Modifiers,
-}
-
-impl Display for KeyPress {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        if self.mods.ctrl() {
-            let _ = f.write_str("Ctrl+");
-        }
-        if self.mods.alt() {
-            let _ = f.write_str("Alt+");
-        }
-        if self.mods.meta() {
-            let _ = f.write_str("Meta+");
-        }
-        if self.mods.shift() {
-            let _ = f.write_str("Shift+");
-        }
-        f.write_str(&self.key.to_string())
-    }
-}
-
-impl KeyPress {
-    pub fn is_char(&self) -> bool {
-        let mut mods = self.mods;
-        mods.set(Modifiers::SHIFT, false);
-        if mods.is_empty() {
-            if let druid::KbKey::Character(_c) = &self.key {
-                return true;
-            }
-        }
-        false
-    }
-
-    pub fn paint(
-        &self,
-        ctx: &mut PaintCtx,
-        origin: Point,
-        config: &Config,
-    ) -> (Point, Vec<(Option<Rect>, PietTextLayout, Point)>) {
-        let mut origin = origin;
-        let mut keys = Vec::new();
-        if self.mods.ctrl() {
-            keys.push("Ctrl".to_string());
-        }
-        if self.mods.alt() {
-            keys.push("Alt".to_string());
-        }
-        if self.mods.meta() {
-            let keyname = match std::env::consts::OS {
-                "macos" => "Cmd",
-                "windows" => "Win",
-                _ => "Meta",
-            };
-            keys.push(keyname.to_string());
-        }
-        if self.mods.shift() {
-            keys.push("Shift".to_string());
-        }
-        match &self.key {
-            druid::keyboard_types::Key::Character(c) => {
-                if *c == c.to_uppercase()
-                    && c.to_lowercase() != c.to_uppercase()
-                    && !self.mods.shift()
-                {
-                    keys.push("Shift".to_string());
-                }
-                keys.push(c.to_uppercase());
-            }
-            _ => {
-                keys.push(self.key.to_string());
-            }
-        }
-
-        let mut items = Vec::new();
-        let keys_len = keys.len();
-        for (i, key) in keys.iter().enumerate() {
-            let (rect, text_layout, text_layout_pos) =
-                paint_key(ctx, key, origin, config);
-            origin += (rect.width() + 5.0, 0.0);
-
-            items.push((Some(rect), text_layout, text_layout_pos));
-
-            if i < keys_len - 1 {
-                let text_layout = ctx
-                    .text()
-                    .new_text_layout("+")
-                    .font(FontFamily::SYSTEM_UI, 13.0)
-                    .text_color(
-                        config
-                            .get_color_unchecked(LapceTheme::EDITOR_FOREGROUND)
-                            .clone(),
-                    )
-                    .build()
-                    .unwrap();
-                let text_size = text_layout.size();
-                let text_layout_pos = origin + (0.0, -(text_size.height / 2.0));
-                items.push((None, text_layout, text_layout_pos));
-                origin += (text_size.width + 5.0, 0.0);
-            }
-        }
-
-        (origin, items)
-    }
 }
 
 pub fn paint_key(
@@ -588,91 +485,7 @@ impl KeyPressData {
                 && self.check_condition(&condition[and_indics[0].0 + 2..], check)
         }
     }
-}
 
-pub struct KeyMapLoader {
-    keymaps: IndexMap<Vec<KeyPress>, Vec<KeyMap>>,
-    command_keymaps: IndexMap<String, Vec<KeyMap>>,
-}
-
-impl KeyMapLoader {
-    pub fn new() -> Self {
-        Self {
-            keymaps: Default::default(),
-            command_keymaps: Default::default(),
-        }
-    }
-
-    pub fn load_from_str<'a>(
-        &'a mut self,
-        s: &str,
-        modal: bool,
-    ) -> Result<&'a mut Self> {
-        let toml_keymaps: toml::Value = toml::from_str(s)?;
-        let toml_keymaps = toml_keymaps
-            .get("keymaps")
-            .and_then(|v| v.as_array())
-            .ok_or_else(|| anyhow!("no keymaps"))?;
-
-        for toml_keymap in toml_keymaps {
-            let keymap = match Self::get_keymap(toml_keymap, modal) {
-                Ok(keymap) => keymap,
-                Err(err) => {
-                    log::error!("Could not parse keymap: {err}");
-                    continue;
-                }
-            };
-
-            let (command, bind) = match keymap.command.strip_prefix('-') {
-                Some(cmd) => (cmd.to_string(), false),
-                None => (keymap.command.clone(), true),
-            };
-
-            let current_keymaps = self.command_keymaps.entry(command).or_default();
-            if bind {
-                current_keymaps.push(keymap.clone());
-                for i in 1..keymap.key.len() + 1 {
-                    let key = keymap.key[..i].to_vec();
-                    self.keymaps.entry(key).or_default().push(keymap.clone());
-                }
-            } else {
-                let is_keymap = |k: &KeyMap| -> bool {
-                    k.when == keymap.when
-                        && k.modes == keymap.modes
-                        && k.key == keymap.key
-                };
-                if let Some(index) = current_keymaps.iter().position(is_keymap) {
-                    current_keymaps.remove(index);
-                }
-                for i in 1..keymap.key.len() + 1 {
-                    if let Some(keymaps) = self.keymaps.get_mut(&keymap.key[..i]) {
-                        if let Some(index) = keymaps.iter().position(is_keymap) {
-                            keymaps.remove(index);
-                        }
-                    }
-                }
-            }
-        }
-
-        Ok(self)
-    }
-
-    pub fn finalize(
-        self,
-    ) -> (
-        IndexMap<Vec<KeyPress>, Vec<KeyMap>>,
-        IndexMap<String, Vec<KeyMap>>,
-    ) {
-        let Self {
-            keymaps: map,
-            command_keymaps: command_map,
-        } = self;
-
-        (map, command_map)
-    }
-}
-
-impl KeyPressData {
     fn get_file_array() -> Option<toml::value::Array> {
         let path = Self::file()?;
         let content = std::fs::read(&path).ok()?;
@@ -884,462 +697,6 @@ impl KeyPressData {
     }
 }
 
-impl KeyPress {
-    /// Convert a piece of text representing a key in a keymaps file to the actual key
-    /// Returns `None` if it was not able to find a key with that name
-    fn map_str_to_key(key_part: &str) -> Option<druid::KbKey> {
-        // Checks if it is a character key
-        fn is_key_string(s: &str) -> bool {
-            s.chars().all(|c| !c.is_control())
-                && s.chars().skip(1).all(|c| !c.is_ascii())
-        }
-
-        // Import into scope to reduce noise
-        use druid::keyboard_types::Key::*;
-        Some(match key_part.to_lowercase().as_str() {
-            s if is_key_string(s) => Character(key_part.to_string()),
-            "unidentified" => Unidentified,
-            "alt" => Alt,
-            "altgraph" => AltGraph,
-            "capslock" => CapsLock,
-            "control" => Control,
-            "fn" => Fn,
-            "fnlock" => FnLock,
-            "meta" => Meta,
-            "numlock" => NumLock,
-            "scrolllock" => ScrollLock,
-            "shift" => Shift,
-            "symbol" => Symbol,
-            "symbollock" => SymbolLock,
-            "hyper" => Hyper,
-            "super" => Super,
-            "enter" => Enter,
-            "tab" => Tab,
-            "arrowdown" => ArrowDown,
-            "arrowleft" => ArrowLeft,
-            "arrowright" => ArrowRight,
-            "arrowup" => ArrowUp,
-            "end" => End,
-            "home" => Home,
-            "pagedown" => PageDown,
-            "pageup" => PageUp,
-            "backspace" => Backspace,
-            "clear" => Clear,
-            "copy" => Copy,
-            "crsel" => CrSel,
-            "cut" => Cut,
-            "delete" => Delete,
-            "eraseeof" => EraseEof,
-            "exsel" => ExSel,
-            "insert" => Insert,
-            "paste" => Paste,
-            "redo" => Redo,
-            "undo" => Undo,
-            "accept" => Accept,
-            "again" => Again,
-            "attn" => Attn,
-            "cancel" => Cancel,
-            "contextmenu" => ContextMenu,
-            "escape" => Escape,
-            "execute" => Execute,
-            "find" => Find,
-            "help" => Help,
-            "pause" => Pause,
-            "play" => Play,
-            "props" => Props,
-            "select" => Select,
-            "zoomin" => ZoomIn,
-            "zoomout" => ZoomOut,
-            "brightnessdown" => BrightnessDown,
-            "brightnessup" => BrightnessUp,
-            "eject" => Eject,
-            "logoff" => LogOff,
-            "power" => Power,
-            "poweroff" => PowerOff,
-            "printscreen" => PrintScreen,
-            "hibernate" => Hibernate,
-            "standby" => Standby,
-            "wakeup" => WakeUp,
-            "allcandidates" => AllCandidates,
-            "alphanumeric" => Alphanumeric,
-            "codeinput" => CodeInput,
-            "compose" => Compose,
-            "convert" => Convert,
-            "dead" => Dead,
-            "finalmode" => FinalMode,
-            "groupfirst" => GroupFirst,
-            "grouplast" => GroupLast,
-            "groupnext" => GroupNext,
-            "groupprevious" => GroupPrevious,
-            "modechange" => ModeChange,
-            "nextcandidate" => NextCandidate,
-            "nonconvert" => NonConvert,
-            "previouscandidate" => PreviousCandidate,
-            "process" => Process,
-            "singlecandidate" => SingleCandidate,
-            "hangulmode" => HangulMode,
-            "hanjamode" => HanjaMode,
-            "junjamode" => JunjaMode,
-            "eisu" => Eisu,
-            "hankaku" => Hankaku,
-            "hiragana" => Hiragana,
-            "hiraganakatakana" => HiraganaKatakana,
-            "kanamode" => KanaMode,
-            "kanjimode" => KanjiMode,
-            "katakana" => Katakana,
-            "romaji" => Romaji,
-            "zenkaku" => Zenkaku,
-            "zenkakuhankaku" => ZenkakuHankaku,
-            "f1" => F1,
-            "f2" => F2,
-            "f3" => F3,
-            "f4" => F4,
-            "f5" => F5,
-            "f6" => F6,
-            "f7" => F7,
-            "f8" => F8,
-            "f9" => F9,
-            "f10" => F10,
-            "f11" => F11,
-            "f12" => F12,
-            "soft1" => Soft1,
-            "soft2" => Soft2,
-            "soft3" => Soft3,
-            "soft4" => Soft4,
-            "channeldown" => ChannelDown,
-            "channelup" => ChannelUp,
-            "close" => Close,
-            "mailforward" => MailForward,
-            "mailreply" => MailReply,
-            "mailsend" => MailSend,
-            "mediaclose" => MediaClose,
-            "mediafastforward" => MediaFastForward,
-            "mediapause" => MediaPause,
-            "mediaplay" => MediaPlay,
-            "mediaplaypause" => MediaPlayPause,
-            "mediarecord" => MediaRecord,
-            "mediarewind" => MediaRewind,
-            "mediastop" => MediaStop,
-            "mediatracknext" => MediaTrackNext,
-            "mediatrackprevious" => MediaTrackPrevious,
-            "new" => New,
-            "open" => Open,
-            "print" => Print,
-            "save" => Save,
-            "spellcheck" => SpellCheck,
-            "key11" => Key11,
-            "key12" => Key12,
-            "audiobalanceleft" => AudioBalanceLeft,
-            "audiobalanceright" => AudioBalanceRight,
-            "audiobassboostdown" => AudioBassBoostDown,
-            "audiobassboosttoggle" => AudioBassBoostToggle,
-            "audiobassboostup" => AudioBassBoostUp,
-            "audiofaderfront" => AudioFaderFront,
-            "audiofaderrear" => AudioFaderRear,
-            "audiosurroundmodenext" => AudioSurroundModeNext,
-            "audiotrebledown" => AudioTrebleDown,
-            "audiotrebleup" => AudioTrebleUp,
-            "audiovolumedown" => AudioVolumeDown,
-            "audiovolumeup" => AudioVolumeUp,
-            "audiovolumemute" => AudioVolumeMute,
-            "microphonetoggle" => MicrophoneToggle,
-            "microphonevolumedown" => MicrophoneVolumeDown,
-            "microphonevolumeup" => MicrophoneVolumeUp,
-            "microphonevolumemute" => MicrophoneVolumeMute,
-            "speechcorrectionlist" => SpeechCorrectionList,
-            "speechinputtoggle" => SpeechInputToggle,
-            "launchapplication1" => LaunchApplication1,
-            "launchapplication2" => LaunchApplication2,
-            "launchcalendar" => LaunchCalendar,
-            "launchcontacts" => LaunchContacts,
-            "launchmail" => LaunchMail,
-            "launchmediaplayer" => LaunchMediaPlayer,
-            "launchmusicplayer" => LaunchMusicPlayer,
-            "launchphone" => LaunchPhone,
-            "launchscreensaver" => LaunchScreenSaver,
-            "launchspreadsheet" => LaunchSpreadsheet,
-            "launchwebbrowser" => LaunchWebBrowser,
-            "launchwebcam" => LaunchWebCam,
-            "launchwordprocessor" => LaunchWordProcessor,
-            "browserback" => BrowserBack,
-            "browserfavorites" => BrowserFavorites,
-            "browserforward" => BrowserForward,
-            "browserhome" => BrowserHome,
-            "browserrefresh" => BrowserRefresh,
-            "browsersearch" => BrowserSearch,
-            "browserstop" => BrowserStop,
-            "appswitch" => AppSwitch,
-            "call" => Call,
-            "camera" => Camera,
-            "camerafocus" => CameraFocus,
-            "endcall" => EndCall,
-            "goback" => GoBack,
-            "gohome" => GoHome,
-            "headsethook" => HeadsetHook,
-            "lastnumberredial" => LastNumberRedial,
-            "notification" => Notification,
-            "mannermode" => MannerMode,
-            "voicedial" => VoiceDial,
-            "tv" => TV,
-            "tv3dmode" => TV3DMode,
-            "tvantennacable" => TVAntennaCable,
-            "tvaudiodescription" => TVAudioDescription,
-            "tvaudiodescriptionmixdown" => TVAudioDescriptionMixDown,
-            "tvaudiodescriptionmixup" => TVAudioDescriptionMixUp,
-            "tvcontentsmenu" => TVContentsMenu,
-            "tvdataservice" => TVDataService,
-            "tvinput" => TVInput,
-            "tvinputcomponent1" => TVInputComponent1,
-            "tvinputcomponent2" => TVInputComponent2,
-            "tvinputcomposite1" => TVInputComposite1,
-            "tvinputcomposite2" => TVInputComposite2,
-            "tvinputhdmi1" => TVInputHDMI1,
-            "tvinputhdmi2" => TVInputHDMI2,
-            "tvinputhdmi3" => TVInputHDMI3,
-            "tvinputhdmi4" => TVInputHDMI4,
-            "tvinputvga1" => TVInputVGA1,
-            "tvmediacontext" => TVMediaContext,
-            "tvnetwork" => TVNetwork,
-            "tvnumberentry" => TVNumberEntry,
-            "tvpower" => TVPower,
-            "tvradioservice" => TVRadioService,
-            "tvsatellite" => TVSatellite,
-            "tvsatellitebs" => TVSatelliteBS,
-            "tvsatellitecs" => TVSatelliteCS,
-            "tvsatellitetoggle" => TVSatelliteToggle,
-            "tvterrestrialanalog" => TVTerrestrialAnalog,
-            "tvterrestrialdigital" => TVTerrestrialDigital,
-            "tvtimer" => TVTimer,
-            "avrinput" => AVRInput,
-            "avrpower" => AVRPower,
-            "colorf0red" => ColorF0Red,
-            "colorf1green" => ColorF1Green,
-            "colorf2yellow" => ColorF2Yellow,
-            "colorf3blue" => ColorF3Blue,
-            "colorf4grey" => ColorF4Grey,
-            "colorf5brown" => ColorF5Brown,
-            "closedcaptiontoggle" => ClosedCaptionToggle,
-            "dimmer" => Dimmer,
-            "displayswap" => DisplaySwap,
-            "dvr" => DVR,
-            "exit" => Exit,
-            "favoriteclear0" => FavoriteClear0,
-            "favoriteclear1" => FavoriteClear1,
-            "favoriteclear2" => FavoriteClear2,
-            "favoriteclear3" => FavoriteClear3,
-            "favoriterecall0" => FavoriteRecall0,
-            "favoriterecall1" => FavoriteRecall1,
-            "favoriterecall2" => FavoriteRecall2,
-            "favoriterecall3" => FavoriteRecall3,
-            "favoritestore0" => FavoriteStore0,
-            "favoritestore1" => FavoriteStore1,
-            "favoritestore2" => FavoriteStore2,
-            "favoritestore3" => FavoriteStore3,
-            "guide" => Guide,
-            "guidenextday" => GuideNextDay,
-            "guidepreviousday" => GuidePreviousDay,
-            "info" => Info,
-            "instantreplay" => InstantReplay,
-            "link" => Link,
-            "listprogram" => ListProgram,
-            "livecontent" => LiveContent,
-            "lock" => Lock,
-            "mediaapps" => MediaApps,
-            "mediaaudiotrack" => MediaAudioTrack,
-            "medialast" => MediaLast,
-            "mediaskipbackward" => MediaSkipBackward,
-            "mediaskipforward" => MediaSkipForward,
-            "mediastepbackward" => MediaStepBackward,
-            "mediastepforward" => MediaStepForward,
-            "mediatopmenu" => MediaTopMenu,
-            "navigatein" => NavigateIn,
-            "navigatenext" => NavigateNext,
-            "navigateout" => NavigateOut,
-            "navigateprevious" => NavigatePrevious,
-            "nextfavoritechannel" => NextFavoriteChannel,
-            "nextuserprofile" => NextUserProfile,
-            "ondemand" => OnDemand,
-            "pairing" => Pairing,
-            "pinpdown" => PinPDown,
-            "pinpmove" => PinPMove,
-            "pinptoggle" => PinPToggle,
-            "pinpup" => PinPUp,
-            "playspeeddown" => PlaySpeedDown,
-            "playspeedreset" => PlaySpeedReset,
-            "playspeedup" => PlaySpeedUp,
-            "randomtoggle" => RandomToggle,
-            "rclowbattery" => RcLowBattery,
-            "recordspeednext" => RecordSpeedNext,
-            "rfbypass" => RfBypass,
-            "scanchannelstoggle" => ScanChannelsToggle,
-            "screenmodenext" => ScreenModeNext,
-            "settings" => Settings,
-            "splitscreentoggle" => SplitScreenToggle,
-            "stbinput" => STBInput,
-            "stbpower" => STBPower,
-            "subtitle" => Subtitle,
-            "teletext" => Teletext,
-            "videomodenext" => VideoModeNext,
-            "wink" => Wink,
-            "zoomtoggle" => ZoomToggle,
-
-            // Custom key name mappings
-            "esc" => Escape,
-            "space" => Character(" ".to_string()),
-            "bs" => Backspace,
-            "up" => ArrowUp,
-            "down" => ArrowDown,
-            "right" => ArrowRight,
-            "left" => ArrowLeft,
-            "del" => Delete,
-
-            _ => return None,
-        })
-    }
-}
-
-impl KeyPress {
-    pub fn parse(key: &str) -> Vec<Self> {
-        key.split(' ')
-            .filter_map(|k| {
-                let (modifiers, key) = match k.rsplit_once('+') {
-                    Some(pair) => pair,
-                    None => ("", k),
-                };
-
-                let key = match Self::map_str_to_key(key) {
-                    Some(key) => key,
-                    None => {
-                        // Skip past unrecognized key definitions
-                        log::warn!("Unrecognized key: {key}");
-                        return None;
-                    }
-                };
-
-                let mut mods = Modifiers::default();
-                for part in modifiers.split('+') {
-                    match part.to_lowercase().as_ref() {
-                        "ctrl" => mods.set(Modifiers::CONTROL, true),
-                        "meta" => mods.set(Modifiers::META, true),
-                        "shift" => mods.set(Modifiers::SHIFT, true),
-                        "alt" => mods.set(Modifiers::ALT, true),
-                        _ => (),
-                    }
-                }
-
-                Some(KeyPress { mods, key })
-            })
-            .collect()
-    }
-}
-
-impl KeyMapLoader {
-    fn get_keymap(toml_keymap: &toml::Value, modal: bool) -> Result<KeyMap> {
-        let key = toml_keymap
-            .get("key")
-            .and_then(|v| v.as_str())
-            .ok_or_else(|| anyhow!("no key in keymap"))?;
-
-        let modes = get_modes(toml_keymap);
-        if !modal && !modes.is_empty() && !modes.contains(&Mode::Insert) {
-            return Err(anyhow!(""));
-        }
-
-        Ok(KeyMap {
-            key: KeyPress::parse(key),
-            modes: get_modes(toml_keymap),
-            when: toml_keymap
-                .get("when")
-                .and_then(|w| w.as_str())
-                .map(|w| w.to_string()),
-            command: toml_keymap
-                .get("command")
-                .and_then(|c| c.as_str())
-                .map(|w| w.trim().to_string())
-                .unwrap_or_else(|| "".to_string()),
-        })
-    }
-}
-
-fn get_modes(toml_keymap: &toml::Value) -> Vec<Mode> {
-    let mut modes = toml_keymap
-        .get("mode")
-        .and_then(|v| v.as_str())
-        .map(|m| {
-            m.chars()
-                .filter_map(|c| match c.to_lowercase().to_string().as_ref() {
-                    "i" => Some(Mode::Insert),
-                    "n" => Some(Mode::Normal),
-                    "v" => Some(Mode::Visual),
-                    "t" => Some(Mode::Terminal),
-                    _ => None,
-                })
-                .collect()
-        })
-        .unwrap_or_else(Vec::new);
-    modes.sort();
-    modes
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_keymap() {
-        let keymaps = r###"
-keymaps = [
-    { key = "ctrl+w l l", command = "right", when = "n" },
-    { key = "ctrl+w l", command = "right", when = "n" },
-    { key = "ctrl+w h", command = "left", when = "n" },
-    { key = "ctrl+w",   command = "left", when = "n" },
-    { key = "End", command = "line_end", when = "n" },
-    { key = "I", command = "insert_first_non_blank", when = "n" },
-]
-        "###;
-        let mut loader = KeyMapLoader::new();
-        loader.load_from_str(keymaps, true).unwrap();
-
-        let (keymaps, _) = loader.finalize();
-
-        // Lower case modifiers
-        let keypress = KeyPress::parse("ctrl+w");
-        assert_eq!(keymaps.get(&keypress).unwrap().len(), 4);
-
-        let keypress = KeyPress::parse("ctrl+w l");
-        assert_eq!(keymaps.get(&keypress).unwrap().len(), 2);
-
-        let keypress = KeyPress::parse("ctrl+w h");
-        assert_eq!(keymaps.get(&keypress).unwrap().len(), 1);
-
-        let keypress = KeyPress::parse("ctrl+w l l");
-        assert_eq!(keymaps.get(&keypress).unwrap().len(), 1);
-
-        let keypress = KeyPress::parse("end");
-        assert_eq!(keymaps.get(&keypress).unwrap().len(), 1);
-
-        // Upper case modifiers
-        let keypress = KeyPress::parse("Ctrl+w");
-        assert_eq!(keymaps.get(&keypress).unwrap().len(), 4);
-
-        let keypress = KeyPress::parse("Ctrl+w l");
-        assert_eq!(keymaps.get(&keypress).unwrap().len(), 2);
-
-        let keypress = KeyPress::parse("Ctrl+w h");
-        assert_eq!(keymaps.get(&keypress).unwrap().len(), 1);
-
-        let keypress = KeyPress::parse("Ctrl+w l l");
-        assert_eq!(keymaps.get(&keypress).unwrap().len(), 1);
-
-        let keypress = KeyPress::parse("End");
-        assert_eq!(keymaps.get(&keypress).unwrap().len(), 1);
-
-        // No modifier
-        let keypress = KeyPress::parse("I");
-        assert_eq!(keymaps.get(&keypress).unwrap().len(), 1);
-    }
-}
-
 pub struct DefaultKeyPressHandler {}
 
 impl KeyPressFocus for DefaultKeyPressHandler {
@@ -1363,4 +720,24 @@ impl KeyPressFocus for DefaultKeyPressHandler {
     }
 
     fn receive_char(&mut self, _ctx: &mut EventCtx, _c: &str) {}
+}
+
+fn get_modes(toml_keymap: &toml::Value) -> Vec<Mode> {
+    let mut modes = toml_keymap
+        .get("mode")
+        .and_then(|v| v.as_str())
+        .map(|m| {
+            m.chars()
+                .filter_map(|c| match c.to_lowercase().to_string().as_ref() {
+                    "i" => Some(Mode::Insert),
+                    "n" => Some(Mode::Normal),
+                    "v" => Some(Mode::Visual),
+                    "t" => Some(Mode::Terminal),
+                    _ => None,
+                })
+                .collect()
+        })
+        .unwrap_or_else(Vec::new);
+    modes.sort();
+    modes
 }
